feat: Use the BaseButtonComponent to manage button UI and functionality (FM-296)
