+++ src/components/buttons/index.ts
@@ -1,12 +1,12 @@
 import CloseButton from './close-button';
-import CloseButtonHtml from './close-button-html'; // Note: Move all code to close-button.ts once the component is fully integrated.
-import CancelButton from './cancel-button';
+import CloseButtonHtml from './close-button/close-button-html'; // Note: Move all code to close-button.ts once the component is fully integrated.
+import CancelButton from './cancel-button/cancel-button';
 import NextButton from './next-button';
-import NoButton from './no-button';
-import PauseButton from './pause-button';
+import NoButton from './no-button/no-button';
+import PauseButton from './pause-button/pause-button';
 import RetryButton from './retry-button';
-import RetryButtonHtml from './retry-button-html'; // Note: Move all code to retry-button.ts once the component is fully integrated.
-import YesButton from './yes-button';
+import RetryButtonHtml from './retry-button/retry-button-html'; // Note: Move all code to retry-button.ts once the component is fully integrated.
+import YesButton from './yes-button/yes-button';
 import PlayButton from './play-button';
 import LevelBloonButton from './level-button';
